Move getLikes and postLike files to likes api folder
